Main (#27)
* feat: Add lottie-react library and update routes for AI and AboutUs page… (#23)

* Developing UI Designs like Figma design

* feat: updates on css style and design system

* feat: updates to css and ui

* Update postcss.config.js, index.html, tailwind.config.js, App.js, App.css, package.json, and index.css

* Update page title and styling in index.html and index.css files

* feat: Refactor translation options UI in Home.js and add mobile view support

* feat: update file paths and styles

* Create hello.yml

* Create FUNDING.yml (#17)

* feat: add sponsarship button

* Add fontawesome and axios dependencies, update routes, and fix styling issues

* Update react-router-dom and @remix-run/router versions

* Add font-awesome icons and social media links to footer, implement login functionality

* feat: Add lottie-react library and update routes for AI and AboutUs pages

---------


* chore: Update header navigation styling and responsiveness (#26)

* feat: update file paths and styles

* Add fontawesome and axios dependencies, update routes, and fix styling issues

* Update react-router-dom and @remix-run/router versions

* Add font-awesome icons and social media links to footer, implement login functionality

* feat: Add lottie-react library and update routes for AI and AboutUs pages

* feat: Add lottie-react library and update routes for AI and AboutUs page… (#23) (#24)

* Developing UI Designs like Figma design

* feat: updates on css style and design system

* feat: updates to css and ui

* Update postcss.config.js, index.html, tailwind.config.js, App.js, App.css, package.json, and index.css

* Update page title and styling in index.html and index.css files

* feat: Refactor translation options UI in Home.js and add mobile view support

* feat: update file paths and styles

* Create hello.yml

* Create FUNDING.yml (#17)

* feat: add sponsarship button

* Add fontawesome and axios dependencies, update routes, and fix styling issues

* Update react-router-dom and @remix-run/router versions

* Add font-awesome icons and social media links to footer, implement login functionality

* feat: Add lottie-react library and update routes for AI and AboutUs pages

---------


* chore: Update header navigation styling and responsiveness

---------


---------
