+++ src/Pages/Components/Header.js
@@ -87,3 +87,26 @@ const Header = () => {
 };
 
 export default Header;
+
+<style jsx>{`
+  .container {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+  }
+
+  nav {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+
+  @media (max-width: 768px) {
+    nav {
+      width: 100%;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+    }
+  }
+`}</style>
